Moved: actionTypes to typescript
